requirements.txt
streamlit
numpy
pydub
librosa
matplotlib
scipy
opencv-python
ffmpeg-python
